# src/auth/dto/LoginDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, IsNotEmpty } from 'class-validator';

export class LoginDTO {
  @ApiProperty({ description: 'User email address' })
  @IsNotEmpty()
  @IsEmail()
  email: string;
  @ApiProperty({ description: 'User password' })
  @IsNotEmpty()
  password: string;
}