# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/colleges/college_data/{college_id}": {
      "get": {
        "operationId": "CollegesController_getCollegeData",
        "parameters": [
          {
            "name": "college_id",
            "required": true,
            "in": "path",
            "description": "The ID of the college",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Average and placement data for the college."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Colleges"
        ]
      }
    },
    "/colleges/college_courses/{college_id}": {
      "get": {
        "operationId": "CollegesController_getCollegeCourses",
        "parameters": [
          {
            "name": "college_id",
            "required": true,
            "in": "path",
            "description": "The ID of the college",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of courses offered by the college sorted by course fee."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Colleges"
        ]
      }
    },
    "/colleges": {
      "get": {
        "operationId": "CollegesController_filterColleges",
        "parameters": [
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "Filter colleges by city name",
            "schema": {
              "example": "New York",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Filter colleges by state name",
            "schema": {
              "example": "California",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered list of colleges based on city or state."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Colleges"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in."
          },
          "401": {
            "description": "Invalid credentials."
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_singup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered."
          },
          "400": {
            "description": "Bad request (e.g. missing or invalid fields)."
          }
        },
        "summary": "User signup",
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "LoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SigninDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      }
    }
  }
}